# todo-app/src/main/java/org/openapitools/api/TodosApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.14.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Todo;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-06-26T13:34:41.538683300+02:00[Europe/Berlin]", comments = "Generator version: 7.14.0")
@Validated
@Tag(name = "todos", description = "the todos API")
public interface TodosApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /todos : Liste aller Todos abrufen
     *
     * @return Erfolgreich geladen (status code 200)
     */
    @Operation(
        operationId = "todosGet",
        summary = "Liste aller Todos abrufen",
        responses = {
            @ApiResponse(responseCode = "200", description = "Erfolgreich geladen", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Todo.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/todos",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Todo>> todosGet(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"id\" : 1, \"title\" : \"Einkaufen\", \"done\" : false }, { \"id\" : 1, \"title\" : \"Einkaufen\", \"done\" : false } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /todos/{id} : Todo löschen
     *
     * @param id  (required)
     * @return Todo gelöscht (status code 204)
     */
    @Operation(
        operationId = "todosIdDelete",
        summary = "Todo löschen",
        responses = {
            @ApiResponse(responseCode = "204", description = "Todo gelöscht")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/todos/{id}"
    )
    
    default ResponseEntity<Void> todosIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /todos/{id}/done : Markiere Todo als erledigt oder nicht erledigt (Toggle)
     * Ändert den Status eines Todos: - Wenn &#x60;done&#x3D;false&#x60;, wird es auf &#x60;true&#x60; gesetzt. - Wenn &#x60;done&#x3D;true&#x60;, wird es auf &#x60;false&#x60; zurückgesetzt. 
     *
     * @param id  (required)
     * @return Erfolgreich aktualisiert (status code 200)
     *         or Todo wurde nicht gefunden (status code 404)
     */
    @Operation(
        operationId = "todosIdDonePut",
        summary = "Markiere Todo als erledigt oder nicht erledigt (Toggle)",
        description = "Ändert den Status eines Todos: - Wenn `done=false`, wird es auf `true` gesetzt. - Wenn `done=true`, wird es auf `false` zurückgesetzt. ",
        responses = {
            @ApiResponse(responseCode = "200", description = "Erfolgreich aktualisiert", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            }),
            @ApiResponse(responseCode = "404", description = "Todo wurde nicht gefunden")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/todos/{id}/done",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Todo> todosIdDonePut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 1, \"title\" : \"Einkaufen\", \"done\" : false }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /todos/{id} : Ein einzelnes Todo abrufen
     *
     * @param id  (required)
     * @return Todo gefunden (status code 200)
     *         or Todo nicht gefunden (status code 404)
     */
    @Operation(
        operationId = "todosIdGet",
        summary = "Ein einzelnes Todo abrufen",
        responses = {
            @ApiResponse(responseCode = "200", description = "Todo gefunden", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            }),
            @ApiResponse(responseCode = "404", description = "Todo nicht gefunden")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/todos/{id}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Todo> todosIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 1, \"title\" : \"Einkaufen\", \"done\" : false }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /todos/{id}/title : Titel eines Todos ändern
     *
     * @param id  (required)
     * @param body  (required)
     * @return Titel geändert (status code 200)
     *         or Todo nicht gefunden (status code 404)
     */
    @Operation(
        operationId = "todosIdTitlePut",
        summary = "Titel eines Todos ändern",
        responses = {
            @ApiResponse(responseCode = "200", description = "Titel geändert", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            }),
            @ApiResponse(responseCode = "404", description = "Todo nicht gefunden")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/todos/{id}/title",
        produces = { "application/json" },
        consumes = { "text/plain" }
    )
    
    default ResponseEntity<Todo> todosIdTitlePut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Integer id,
        @Parameter(name = "body", description = "", required = true) @Valid @RequestBody String body
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 1, \"title\" : \"Einkaufen\", \"done\" : false }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /todos : Ein neues Todo hinzufügen
     *
     * @param todo  (required)
     * @return Todo erfolgreich erstellt (status code 201)
     */
    @Operation(
        operationId = "todosPost",
        summary = "Ein neues Todo hinzufügen",
        responses = {
            @ApiResponse(responseCode = "201", description = "Todo erfolgreich erstellt", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Todo.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/todos",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Todo> todosPost(
        @Parameter(name = "Todo", description = "", required = true) @Valid @RequestBody Todo todo
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"id\" : 1, \"title\" : \"Einkaufen\", \"done\" : false }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}